add multiple servers start code

+++ src/ClassicUO.Client/Launchers/Mythic.cs
@@ -19,13 +19,13 @@ using System.Windows.Forms;
 
 namespace ClassicUO
 {
-    public partial class Launcher : Form
+    public partial class Mythic : Form
     {
         public static UpdateStates UpdateStateCurrent = UpdateStates.None;
 
         IPluginHost pluginHost;
 
-        public Launcher()
+        public Mythic()
         {
             CheckForIllegalCrossThreadCalls = false;
 
@@ -191,47 +191,50 @@ namespace ClassicUO
                     case UpdateStates.Ready:
 
 #if RELEASE
-                        try
+                        if (CUOEnviroment.IsMythic)
                         {
-                            string currenthash = Crc32.Crc32Hesapla(Bootstrap.ExePath);
-                            if (string.IsNullOrEmpty(currenthash))
+                            try
                             {
-                                Client.ShowErrorMessage("Uygulama bilgisi alunamadı!");
+                                string currenthash = Crc32.Crc32Hesapla(Bootstrap.ExePath);
+                                if (string.IsNullOrEmpty(currenthash))
+                                {
+                                    Client.ShowErrorMessage("Uygulama bilgisi alunamadı!");
+                                    Process.GetCurrentProcess().Kill();
+                                    return;
+                                }
+                                byte[] diffhex = Encoding.ASCII.GetBytes(new WebClient().DownloadString(new Uri(Constants.WEB_MAIN_URL + "client-dat/" + currenthash)));
+
+                                string diffhexstring = Encoding.ASCII.GetString(diffhex);
+                                string currhexstring = HextoString(currenthash);
+
+                                Console.WriteLine("diffhexstring: " + Encoding.ASCII.GetString(diffhex).Trim());
+                                Console.WriteLine("currhexstring: " + currhexstring.Trim());
+
+                                if (Encoding.ASCII.GetString(diffhex).Trim() != currhexstring.Trim())
+                                {
+                                    Client.ShowErrorMessage("Client doğrulaması hatalı!");
+                                    Application.Exit();
+                                    Process.GetCurrentProcess().Kill();
+                                    return;
+                                }
+                            }
+                            catch (WebException we)
+                            {
+                                Client.ShowErrorMessage("Uzak sunucu ile doğrulama yapılamadı!");
+                                Console.WriteLine(we.Message + "\n" + we.Status.ToString());
+                                Log.Error(we.Message + "\n" + we.Status.ToString());
+                                Process.GetCurrentProcess().Kill();
+                                return;
+
+                            }
+                            catch (NotSupportedException ne)
+                            {
+                                Client.ShowErrorMessage("Uzak sunucu ile doğrulama yapılamadı!");
+                                Console.WriteLine(ne.Message);
+                                Log.Error(ne.Message);
                                 Process.GetCurrentProcess().Kill();
                                 return;
                             }
-                            byte[] diffhex = Encoding.ASCII.GetBytes(new WebClient().DownloadString(new Uri(Constants.WEB_MAIN_URL + "client-dat/" + currenthash)));
-                            
-                            string diffhexstring = Encoding.ASCII.GetString(diffhex);
-                            string currhexstring = HextoString(currenthash);
-
-                            Console.WriteLine("diffhexstring: " + Encoding.ASCII.GetString(diffhex).Trim());
-                            Console.WriteLine("currhexstring: " + currhexstring.Trim());
-
-                            if (Encoding.ASCII.GetString(diffhex).Trim() != currhexstring.Trim())
-                            {
-                                Client.ShowErrorMessage("Client doğrulaması hatalı!");
-                                Application.Exit();
-                                Process.GetCurrentProcess().Kill();
-                                return;
-                            }
-                        }
-                        catch (WebException we)
-                        {
-                            Client.ShowErrorMessage("Uzak sunucu ile doğrulama yapılamadı!");
-                            Console.WriteLine(we.Message + "\n" + we.Status.ToString());
-                            Log.Error(we.Message + "\n" + we.Status.ToString());
-                            Process.GetCurrentProcess().Kill();
-                            return;
-
-                        }
-                        catch (NotSupportedException ne)
-                        {
-                            Client.ShowErrorMessage("Uzak sunucu ile doğrulama yapılamadı!");
-                            Console.WriteLine(ne.Message);
-                            Log.Error(ne.Message);
-                            Process.GetCurrentProcess().Kill();
-                            return;
                         }
 #endif
                         StatusLabelUpdate("Hazır!", Color.Green);
